- Fee module: Base/max rates, stress-based adjustments, decay times   - Liquidity module: Mean-reverting SDEs with jumps, provision rates   - Hawkes module: Self-exciting processes with stability validation (α < 1)
  Locally, 29 tests passing (20 module tests + 9 utility fixes).

+++ ethode/runtime.py
@@ -40,14 +40,14 @@ class QuantityNode(struct.Struct):
         cls,
         value: float,
         units: UnitSpec,
-        dtype: jnp.dtype = jnp.float32
+        dtype: jnp.dtype = jnp.float64
     ) -> QuantityNode:
         """Create a QuantityNode from a float value.
 
         Args:
             value: Numerical value in canonical units
             units: Unit specification
-            dtype: JAX array dtype (default float32)
+            dtype: JAX array dtype (default float64)
 
         Returns:
             QuantityNode instance
